+++ hnn_core/__init__.py
@@ -1,10 +1,11 @@
 from .dipole import simulate_dipole, read_dipole, average_dipoles, Dipole
 from .params import Params, read_params
-from .network import Network, pick_connection, read_network
+from .network import Network, pick_connection
 from .network_models import jones_2009_model, law_2021_model, calcium_model
 from .cell import Cell
 from .cell_response import CellResponse, read_spikes
 from .cells_default import pyramidal, basket
 from .parallel_backends import MPIBackend, JoblibBackend
+from .io import read_network
 
 __version__ = '0.4.dev0'
