+++ hnn_core/network.py
@@ -10,14 +10,10 @@
 import itertools as it
 from copy import deepcopy
 from collections import OrderedDict
-import os
 
 import numpy as np
 import warnings
-from h5io import write_hdf5, read_hdf5
 
-from .cell import Cell, Section
-from .cell_response import CellResponse
 from .drives import _drive_cell_event_times
 from .drives import _get_target_properties, _add_drives_from_params
 from .drives import _check_drive_parameter_values, _check_poisson_rates
@@ -27,6 +23,7 @@ from .viz import plot_cells
 from .externals.mne import _validate_type, _check_option
 from .extracellular import ExtracellularArray
 from .check import _check_gids, _gid_to_type, _string_input_to_list
+from .io import write_network
 
 
 def _create_cell_coords(n_pyr_x, n_pyr_y, zdiff, inplane_distance):
@@ -1429,308 +1426,7 @@ class Network(object):
         return plot_cells(net=self, ax=ax, show=show)
 
     def write(self, fname, overwrite=True, save_unsimulated=False):
-        # Tasks
-        # 1) Remove params dict (Partial done)
-        # 2) Add overwrite warnings (Done)
-        # 3) Add boolean to read/write un-simulated network (Done)
-        # 4) Add boolean to read/write simulated network (Done)
-        # 5) Think about documentation
-        # 6) Add tests
-        # 7) Store the type of object saved and issue warning (Done)
-        if overwrite is False and os.path.exists(fname):
-            raise FileExistsError('File already exists at path %s. Rename '
-                                  'the file or set overwrite=True.' % (fname,))
-        net_data = dict()
-        net_data['object_type'] = "Network"
-        # Write the required params
-        params = dict()
-        params['N_pyr_x'] = self._N_pyr_x
-        params['N_pyr_y'] = self._N_pyr_y
-        params['threshold'] = self.threshold
-        params['celsius'] = self._params['celsius']
-        net_data['params'] = params
-        cell_types_data = dict()
-        for key in self.cell_types:
-            cell_copy = self.cell_types[key].copy()
-            cell_copy.build()
-            cell_types_data[key] = cell_copy.to_dict()
-        net_data['cell_types'] = cell_types_data
-        # Write gid_ranges
-        gid_ranges_data = dict()
-        for key in self.gid_ranges:
-            gid_ranges_data[key] = dict()
-            gid_ranges_data[key]['start'] = self.gid_ranges[key].start
-            gid_ranges_data[key]['stop'] = self.gid_ranges[key].stop
-        net_data['gid_ranges'] = gid_ranges_data
-        # Write pos_dict
-        pos_dict_data = dict()
-        for key in self.pos_dict:
-            pos_dict_data[key] = self.pos_dict[key]
-        net_data['pos_dict'] = pos_dict_data
-        # Write cell_response
-        net_data['cell_response'] = (_cell_response_to_dict
-                                     (self.cell_response, save_unsimulated))
-        # Write External drives
-        external_drives_data = dict()
-        for key in self.external_drives.keys():
-            external_drives_data[key] = (_external_drive_to_dict
-                                         (self.external_drives[key],
-                                          save_unsimulated))
-        net_data['external_drives'] = external_drives_data
-        # Write External biases
-        net_data['external_biases'] = self.external_biases
-        # Write connectivity
-        net_data['connectivity'] = _write_connectivity(self.connectivity)
-        # Write rec arrays
-        net_data['rec_arrays'] = dict()
-        for key in self.rec_arrays.keys():
-            rec_array_copy = self.rec_arrays[key].copy()
-            if save_unsimulated:
-                rec_array_copy._reset()
-            net_data['rec_arrays'][key] = rec_array_copy.to_dict()
-        # Write threshold
-        net_data['threshold'] = self.threshold
-        # Write delay
-        net_data['delay'] = self.delay
-
-        # Saving file
-        write_hdf5(fname, net_data, overwrite=overwrite)
-
-
-def _write_connectivity(connectivity):
-    conns_data = list()
-    for conn in connectivity:
-        conn_data = dict()
-        conn_data['target_type'] = conn['target_type']
-        conn_data['target_gids'] = list(conn['target_gids'])
-        conn_data['num_targets'] = conn['num_targets']
-        conn_data['src_type'] = conn['src_type']
-        conn_data['src_gids'] = list(conn['src_gids'])
-        conn_data['num_srcs'] = conn['num_srcs']
-        gid_pairs = (dict((str(key), val)
-                     for key, val in conn['gid_pairs'].items()))
-        conn_data['gid_pairs'] = gid_pairs
-        conn_data['loc'] = conn['loc']
-        conn_data['receptor'] = conn['receptor']
-        conn_data['nc_dict'] = conn['nc_dict']
-        conn_data['allow_autapses'] = int(conn['allow_autapses'])
-        conn_data['probability'] = conn['probability']
-        conns_data.append(conn_data)
-    return conns_data
-
-
-def _cell_response_to_dict(cell_response, save_unsimulated):
-    # Will be deleted when to_dict method will be merged
-    # in a diiferent PR
-    if (not cell_response) or save_unsimulated:
-        return None
-    cell_response_data = dict()
-    cell_response_data['spike_times'] = cell_response.spike_times
-    cell_response_data['spike_gids'] = cell_response.spike_gids
-    cell_response_data['spike_types'] = cell_response.spike_types
-    vsec_data = cell_response.vsec
-    cell_response_data['vsec'] = list()
-    for trial in vsec_data:
-        trial = dict((str(key), val) for key, val in trial.items())
-        cell_response_data['vsec'].append(trial)
-    isec_data = cell_response.isec
-    cell_response_data['isec'] = list()
-    for trial in isec_data:
-        trial = dict((str(key), val) for key, val in trial.items())
-        cell_response_data['isec'].append(trial)
-    cell_response_data['times'] = cell_response.times
-    return cell_response_data
-
-
-def _external_drive_to_dict(drive, save_unsimulated):
-    drive_data = dict()
-    for key in drive.keys():
-        # Cannot store sets with hdf5
-        if isinstance(drive[key], set):
-            drive_data[key] = list(drive[key])
-        else:
-            drive_data[key] = drive[key]
-    if save_unsimulated:
-        drive_data['events'] = list()
-    return drive_data
-
-
-def _read_cell_types(cell_types_data):
-    cell_types = dict()
-    for cell_name in cell_types_data:
-        cell_data = cell_types_data[cell_name]
-        sections = dict()
-        sections_data = cell_data['sections']
-        for section_name in sections_data:
-            section_data = sections_data[section_name]
-            sections[section_name] = Section(L=section_data['L'],
-                                             diam=section_data['diam'],
-                                             cm=section_data['cm'],
-                                             Ra=section_data['Ra'],
-                                             end_pts=section_data['end_pts'])
-            # Set section attributes
-            sections[section_name].syns = section_data['syns']
-            # Think how to set mechs
-            sections[section_name].mechs = section_data['mechs']
-        cell_types[cell_name] = Cell(name=cell_data['name'],
-                                     pos=cell_data['pos'],
-                                     sections=sections,
-                                     synapses=cell_data['synapses'],
-                                     topology=cell_data['topology'],
-                                     sect_loc=cell_data['sect_loc'],
-                                     gid=cell_data['gid'])
-        # Setting cell attributes
-        cell_types[cell_name].dipole_pp = cell_data['dipole_pp']
-        cell_types[cell_name].vsec = cell_data['vsec']
-        cell_types[cell_name].isec = cell_data['isec']
-        cell_types[cell_name].tonic_biases = cell_data['tonic_biases']
-
-    return cell_types
-
-
-def _read_cell_response(cell_response_data, read_raw):
-    if (not cell_response_data) or read_raw:
-        return None
-    cell_response = CellResponse(spike_times=cell_response_data['spike_times'],
-                                 spike_gids=cell_response_data['spike_gids'],
-                                 spike_types=cell_response_data['spike_types'])
-
-    cell_response._times = cell_response_data['times']
-    cell_response._vsec = list()
-    for trial in cell_response_data['vsec']:
-        trial = dict((int(key), val) for key, val in trial.items())
-        cell_response._vsec.append(trial)
-    cell_response._isec = list()
-    for trial in cell_response_data['isec']:
-        trial = dict((int(key), val) for key, val in trial.items())
-        cell_response._isec.append(trial)
-    return cell_response
-
-
-def _read_external_drive(net, drive_data, read_raw):
-    if drive_data['type'] == 'evoked':
-        # Skipped n_drive_cells here
-        net.add_evoked_drive(name=drive_data['name'],
-                             mu=drive_data['dynamics']['mu'],
-                             sigma=drive_data['dynamics']['sigma'],
-                             numspikes=drive_data['dynamics']['numspikes'],
-                             location=drive_data['location'],
-                             cell_specific=drive_data['cell_specific'],
-                             weights_ampa=drive_data['weights_ampa'],
-                             weights_nmda=drive_data['weights_nmda'],
-                             synaptic_delays=drive_data['synaptic_delays'],
-                             event_seed=drive_data['event_seed'],
-                             conn_seed=drive_data['conn_seed'])
-    elif drive_data['type'] == 'poisson':
-        net.add_poisson_drive(name=drive_data['name'],
-                              tstart=drive_data['dynamics']['tstart'],
-                              tstop=drive_data['dynamics']['tstop'],
-                              rate_constant=(drive_data['dynamics']
-                                                       ['rate_constant']),
-                              location=drive_data['location'],
-                              n_drive_cells=drive_data['n_drive_cells'],
-                              cell_specific=drive_data['cell_specific'],
-                              weights_ampa=drive_data['weights_ampa'],
-                              weights_nmda=drive_data['weights_nmda'],
-                              synaptic_delays=drive_data['synaptic_delays'],
-                              event_seed=drive_data['event_seed'],
-                              conn_seed=drive_data['conn_seed'])
-    elif drive_data['type'] == 'bursty':
-        net.add_bursty_drive(name=drive_data['name'],
-                             tstart=drive_data['dynamics']['tstart'],
-                             tstart_std=drive_data['dynamics']['tstart_std'],
-                             tstop=drive_data['dynamics']['tstop'],
-                             burst_rate=drive_data['dynamics']['burst_rate'],
-                             burst_std=drive_data['dynamics']['burst_std'],
-                             num_spikes=drive_data['dynamics']['num_spikes'],
-                             spike_isi=drive_data['dynamics']['spike_isi'],
-                             location=drive_data['location'],
-                             n_drive_cells=drive_data['n_drive_cells'],
-                             cell_specific=drive_data['cell_specific'],
-                             weights_ampa=drive_data['weights_ampa'],
-                             weights_nmda=drive_data['weights_nmda'],
-                             synaptic_delays=drive_data['synaptic_delays'],
-                             event_seed=drive_data['event_seed'],
-                             conn_seed=drive_data['conn_seed'])
-
-    net.external_drives[drive_data['name']]['events'] = drive_data['events']
-    if read_raw:
-        net.external_drives[drive_data['name']]['events'] = list()
-
-
-def _read_connectivity(net, conns_data):
-    # Overwrite drive connections
-    net.connectivity = list()
-    for conn_data in conns_data:
-        conn_data['allow_autapses'] = bool(conn_data['allow_autapses'])
-        # conn_data['allow_autapses'] = bool(conn_data['allow_autapses'])
-        net.add_connection(src_gids=conn_data['src_type'],
-                           target_gids=conn_data['target_type'],
-                           loc=conn_data['loc'],
-                           receptor=conn_data['receptor'],
-                           weight=conn_data['nc_dict']['A_weight'],
-                           delay=conn_data['nc_dict']['A_delay'],
-                           lamtha=conn_data['nc_dict']['lamtha'],
-                           allow_autapses=conn_data['allow_autapses'],
-                           probability=conn_data['probability'])
-
-
-def _read_rec_arrays(net, rec_arrays_data, read_raw):
-    for key in rec_arrays_data:
-        rec_array = rec_arrays_data[key]
-        net.add_electrode_array(name=key,
-                                electrode_pos=rec_array['positions'],
-                                conductivity=rec_array['conductivity'],
-                                method=rec_array['method'],
-                                min_distance=rec_array['min_distance'])
-        net.rec_arrays[key]._times = rec_array['times']
-        net.rec_arrays[key]._data = rec_array['voltages']
-        if read_raw:
-            net.rec_arrays[key]._reset()
-
-
-def read_network(fname, read_raw=False):
-    net_data = read_hdf5(fname)
-    if 'object_type' not in net_data:
-        raise NameError('The given file is not compatible. '
-                        'The file should contain information'
-                        ' about object type to be read.')
-    if net_data['object_type'] != 'Network':
-        raise ValueError('The object should be of type Network. '
-                         'The file contains object of '
-                         'type %s' % (net_data['object_type'],))
-    params = net_data['params']
-    net = Network(params)
-    # Setting attributes
-    # Set cell types
-    net.cell_types = _read_cell_types(net_data['cell_types'])
-    # Set gid ranges
-    gid_ranges_data = dict()
-    for key in net_data['gid_ranges']:
-        start = net_data['gid_ranges'][key]['start']
-        stop = net_data['gid_ranges'][key]['stop']
-        gid_ranges_data[key] = range(start, stop)
-    net.gid_ranges = gid_ranges_data
-    # Set pos_dict
-    net.pos_dict = net_data['pos_dict']
-    # Set cell_response
-    net.cell_response = _read_cell_response(net_data['cell_response'],
-                                            read_raw)
-    # Set external drives
-    for key in net_data['external_drives'].keys():
-        _read_external_drive(net, net_data['external_drives'][key], read_raw)
-    # Set external biases
-    net.external_biases = net_data['external_biases']
-    # Set connectivity
-    _read_connectivity(net, net_data['connectivity'])
-    # Set rec_arrays
-    _read_rec_arrays(net, net_data['rec_arrays'], read_raw)
-    # Set threshold
-    net.threshold = net_data['threshold']
-    # Set delay
-    net.delay = net_data['delay']
-
-    return net
+        write_network(self, fname, overwrite, save_unsimulated)
 
 
 class _Connectivity(dict):
